+++ core/core.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"net/http"
 	"strconv"
-	"strings"
 	"time"
 	"unsafe"
 
@@ -22,34 +21,6 @@ type options struct {
 	httpClient HttpClient
 }
 
-// DiagramGraph defines the diagram graph.
-type DiagramGraph struct {
-	Title  string  `json:"title"`
-	Footer string  `json:"footer"`
-	Nodes  []*Node `json:"nodes"`
-	Links  []*Link `json:"links"`
-}
-
-// Node diagram's definition node.
-type Node struct {
-	ID         string `json:"id"`
-	Label      string `json:"label"`
-	Group      string `json:"group"`
-	Technology string `json:"technology"`
-	External   bool   `json:"external"`
-	IsQueue    bool   `json:"is_queue"`
-	IsDatabase bool   `json:"is_database"`
-}
-
-// Link diagram's definition link.
-type Link struct {
-	From       string `json:"from"`
-	To         string `json:"to"`
-	Direction  string `json:"direction"`
-	Label      string `json:"label"`
-	Technology string `json:"technology"`
-}
-
 // ClientPlantUML client to communicate with the plantuml server.
 type ClientPlantUML interface {
 	// GenerateDiagram generates the SVG diagram using the diagram as code input.
@@ -60,50 +31,6 @@ type client struct {
 	options options
 
 	baseURL string
-}
-
-func trimQuotes(s string) string {
-	if s[:1] == `"` {
-		s = s[1:]
-	}
-	if s[len(s)-1:] == `"` {
-		s = s[:len(s)-1]
-	}
-	return s
-}
-
-func diagramGraph2Footer(graph DiagramGraph) string {
-	if graph.Footer == "" {
-		return `footer "generated by diagramastext.dev - %date('yyyy-MM-dd')"`
-	}
-	return `footer "` + stringCleaner(graph.Footer) + `"`
-}
-
-func stringCleaner(s string) string {
-	s = strings.TrimSpace(s)
-	s = trimQuotes(s)
-	s = strings.ReplaceAll(s, "\n", "\\n")
-	return s
-}
-
-// diagramGraph2plantUMLCode function to "transpile" the diagram definition graph to plantUML code as string.
-func diagramGraph2plantUMLCode(graph DiagramGraph) string {
-	o := `@startuml
-!include https://raw.githubusercontent.com/plantuml-stdlib/C4-PlantUML/master/C4_Container.puml`
-
-	o += "\n" + diagramGraph2Footer(graph)
-
-	if graph.Title != "" {
-		o += "\n" + diagramGraph2Title(graph)
-	}
-
-	o += "\n@enduml"
-
-	return o
-}
-
-func diagramGraph2Title(graph DiagramGraph) string {
-	return `title "` + stringCleaner(graph.Title) + `"`
 }
 
 func (c *client) GenerateDiagram(v DiagramGraph) ([]byte, error) {
