+++ core/core_test.go
@@ -138,62 +138,3 @@ bar`),
 		)
 	}
 }
-
-func Test_diagramGraph2plantUMLCode(t *testing.T) {
-	type args struct {
-		graph DiagramGraph
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "default graph",
-			args: args{
-				graph: DiagramGraph{},
-			},
-			want: `@startuml
-!include https://raw.githubusercontent.com/plantuml-stdlib/C4-PlantUML/master/C4_Container.puml
-footer "generated by diagramastext.dev - %date('yyyy-MM-dd')"
-@enduml`,
-		},
-		{
-			name: "graph: custom footer",
-			args: args{
-				graph: DiagramGraph{
-					Footer: `  foobar
-"bazqux
-quxx"  `,
-				},
-			},
-			want: `@startuml
-!include https://raw.githubusercontent.com/plantuml-stdlib/C4-PlantUML/master/C4_Container.puml
-footer "foobar\n"bazqux\nquxx"
-@enduml`,
-		},
-		{
-			name: "graph: custom footer and title",
-			args: args{
-				graph: DiagramGraph{
-					Title:  "foo",
-					Footer: "bar",
-				},
-			},
-			want: `@startuml
-!include https://raw.githubusercontent.com/plantuml-stdlib/C4-PlantUML/master/C4_Container.puml
-footer "bar"
-title "foo"
-@enduml`,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(
-			tt.name, func(t *testing.T) {
-				if got := diagramGraph2plantUMLCode(tt.args.graph); got != tt.want {
-					t.Errorf("diagramGraph2plantUMLCode() = %v, want %v", got, tt.want)
-				}
-			},
-		)
-	}
-}
